langchain-basics/pyproject.toml
[project]
name = "langchain-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain-community>=0.4.1",
    "langchain[mistralai,openai]>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
